aula 1.9

+++ src/main/kotlin/br/com/alura/alugames/modelo/InfoJogoJson.kt
@@ -1,10 +1,8 @@
 package br.com.alura.alugames.modelo
 
-import java.math.BigDecimal
-
 data class InfoJogoJson(
     val titulo: String,
     val capa: String,
-    val preco: BigDecimal,
+    val preco: Double,
     val descricao: String
 )
